Use byte for subscription ID
Instead of integer. This is to be consistent with the introduction of a
similar ID for publishers (which will be a byte as well).

+++ deps/rabbitmq_stream/test/rabbit_stream_SUITE_data/src/test/java/com/rabbitmq/stream/StreamTest.java
@@ -87,7 +87,7 @@ public class StreamTest {
                 })
         );
 
-        consumer.subscribe(1, stream, OffsetSpecification.first(), 10);
+        consumer.subscribe((byte) 1, stream, OffsetSpecification.first(), 10);
 
         assertThat(consumingLatch.await(10, TimeUnit.SECONDS)).isTrue();
         assertThat(bodies).hasSize(messageCount);
